move services inside components folder
