+++ apps/web/src/app/simulator/page.tsx
@@ -1,4 +1,4 @@
-import ParcelSimulatorService from "@/services/ParcelSimulator"
+import ParcelSimulatorService from "@/components/services/ParcelSimulator"
 
 export default function SimulatorPage() {
   return <ParcelSimulatorService />
