+++ apps/web/src/components/services/ParcelSimulator/ParcelSimulatorCards/ParcelSimulatorCards.styled.tsx
