+++ apps/web/src/components/services/ParcelSimulator/ParcelSimulatorResult/ParcelSimulatorResult.styled.tsx
