+++ apps/web/src/components/services/ParcelSimulator/ParcelSimulatorResult/index.tsx
